Added local-storage related functions in utility module
