+++ ocs_ci/ocs/resources/storage_cluster.py
@@ -1,22 +1,16 @@
 """
 StorageCluster related functionalities
 """
-import os
-import shutil
-import yaml
-import json
 from ocs_ci.ocs.ocp import OCP, get_images
 from jsonschema import validate
 from ocs_ci.framework import config
 import logging
 from ocs_ci.ocs import constants, defaults, ocp
-from ocs_ci.ocs.resources import csv
-
 from ocs_ci.ocs.resources.csv import CSV
 from ocs_ci.ocs.resources.packagemanifest import get_selector_for_ocs_operator, PackageManifest
-from ocs_ci.utility import utils
-from ocs_ci.ocs.node import get_typed_nodes, get_compute_node_names
-from ocs_ci.utility.retry import retry
+from ocs_ci.ocs.node import get_compute_node_names
+from ocs_ci.utility import utils, localstorage
+
 
 log = logging.getLogger(__name__)
 
@@ -401,20 +395,20 @@ def add_capacity(osd_size_capacity_requested):
     device_sets_required = int(osd_size_capacity_requested / osd_size_existing)
     old_storage_devices_sets_count = get_deviceset_count()
     new_storage_devices_sets_count = int(device_sets_required + old_storage_devices_sets_count)
-    lvpresent = check_local_volume()
+    lvpresent = localstorage.check_local_volume()
     if lvpresent:
-        final_device_list = get_new_device_paths(device_sets_required, osd_size_capacity_requested)
+        final_device_list = localstorage.get_new_device_paths(device_sets_required, osd_size_capacity_requested)
         param = f"""[{{ "op": "replace", "path": "/spec/storageClassDevices/0/devicePaths",
                                                  "value": {final_device_list}}}]"""
         log.info(f"Final device list : {final_device_list}")
-        lvcr = get_local_volume_cr()
+        lvcr = localstorage.get_local_volume_cr()
         log.info("Patching Local Volume CR...")
         lvcr.patch(
             resource_name=lvcr.get()['items'][0]['metadata']['name'],
             params=param.strip('\n'),
             format_type='json'
         )
-        check_pvs_created(int(len(final_device_list) / new_storage_devices_sets_count))
+        localstorage.check_pvs_created(int(len(final_device_list) / new_storage_devices_sets_count))
     sc = get_storage_cluster()
     # adding the storage capacity to the cluster
     params = f"""[{{ "op": "replace", "path": "/spec/storageDeviceSets/0/count",
@@ -493,122 +487,3 @@ def get_all_storageclass():
         )
     ]
     return storageclass
-
-
-def get_local_volume_cr():
-    """
-    Get localVolumeCR object
-
-    Returns:
-        local volume (obj): Local Volume object handler
-
-    """
-    ocp_obj = OCP(kind=constants.LOCAL_VOLUME, namespace=constants.LOCAL_STORAGE_NAMESPACE)
-    return ocp_obj
-
-
-def get_new_device_paths(device_sets_required, osd_size_capacity_requested):
-    """
-    Get new device paths to add capacity over Baremetal cluster
-
-    Args:
-        device_sets_required (int) : Count of device sets to be added
-        osd_size_capacity_requested (int) : Requested OSD size capacity
-
-    Returns:
-        list : List containing added device paths
-
-    """
-    ocp_obj = OCP()
-    workers = get_typed_nodes(node_type="worker")
-    worker_names = [worker.name for worker in workers]
-    output = ocp_obj.exec_oc_cmd("get localvolume local-block -n local-storage -o yaml")
-    cur_device_list = output["spec"]["storageClassDevices"][0]["devicePaths"]
-    path = os.path.join(constants.EXTERNAL_DIR, "device-by-id-ocp")
-    utils.clone_repo(constants.OCP_QE_DEVICEPATH_REPO, path)
-    os.chdir(path)
-    utils.run_cmd("ansible-playbook devices_by_id.yml")
-    with open("local-storage-block.yaml", "r") as cloned_file:
-        with open("local-block.yaml", "w") as our_file:
-            device_from_worker1 = device_sets_required
-            device_from_worker2 = device_sets_required
-            device_from_worker3 = device_sets_required
-            cur_line = cloned_file.readline()
-            while "devicePaths:" not in cur_line:
-                our_file.write(cur_line)
-                cur_line = cloned_file.readline()
-            our_file.write(cur_line)
-            cur_line = cloned_file.readline()
-            # Add required number of device path from each node
-            while cur_line:
-                if str(osd_size_capacity_requested) in cur_line:
-                    if device_from_worker1 and (str(worker_names[0]) in cur_line):
-                        if not any(s in cur_line for s in cur_device_list):
-                            our_file.write(cur_line)
-                            device_from_worker1 = device_from_worker1 - 1
-                    if device_from_worker2 and (str(worker_names[1]) in cur_line):
-                        if not any(s in cur_line for s in cur_device_list):
-                            our_file.write(cur_line)
-                            device_from_worker2 = device_from_worker2 - 1
-                    if device_from_worker3 and (str(worker_names[2]) in cur_line):
-                        if not any(s in cur_line for s in cur_device_list):
-                            our_file.write(cur_line)
-                            device_from_worker3 = device_from_worker3 - 1
-                cur_line = cloned_file.readline()
-    local_block_yaml = open("local-block.yaml")
-    lvcr = yaml.load(local_block_yaml, Loader=yaml.FullLoader)
-    new_dev_paths = lvcr["spec"]["storageClassDevices"][0]["devicePaths"]
-    log.info(f"Newly added devices are: {new_dev_paths}")
-    assert len(new_dev_paths) == (len(worker_names) * device_sets_required), (
-        f"Current devices available = {len(new_dev_paths)}"
-    )
-    os.chdir(constants.TOP_DIR)
-    shutil.rmtree(path)
-    cur_device_list.extend(new_dev_paths)
-    return cur_device_list
-
-
-def check_local_volume():
-    """
-    Function to check if Local-volume is present or not
-
-    Returns:
-        bool: True if LV present, False if LV not present
-
-    """
-
-    if csv.get_csvs_start_with_prefix(
-        csv_prefix=defaults.LOCAL_STORAGE_OPERATOR_NAME,
-        namespace=constants.LOCAL_STORAGE_NAMESPACE
-    ):
-        ocp_obj = OCP()
-        command = "get localvolume local-block -n local-storage "
-        status = ocp_obj.exec_oc_cmd(command, out_yaml_format=False)
-        return "No resources found" not in status
-
-
-@retry(AssertionError, 12, 10, 1)
-def check_pvs_created(num_pvs_required):
-    """
-    Verify that exact number of PVs were created and are in the Available state
-
-    Args:
-        num_pvs_required (int): number of PVs required
-
-    Raises:
-        AssertionError: if the number of PVs are not in the Available state
-
-    """
-    log.info("Verifying PVs are created")
-    out = utils.run_cmd("oc get pv -o json")
-    pv_json = json.loads(out)
-    current_count = 0
-    for pv in pv_json['items']:
-        pv_state = pv['status']['phase']
-        pv_name = pv['metadata']['name']
-        log.info("%s is %s", pv_name, pv_state)
-        if pv_state == 'Available':
-            current_count = current_count + 1
-    assert current_count >= num_pvs_required, (
-        f"Current Available PV count is {current_count}"
-    )
